requirements.txt
google-generativeai>=0.7.0
ollama>=0.3.0
pdfplumber>=0.11.0
python-docx>=1.1.0
